PLUS: Minor fixes from upstream
Remove incorrect cast
MemAlloc() takes a size_t in Plus, but an int in ScottFree, which is inconsistent and causes
things like this.

Fix a few small type mismatches
uint64_t isn't guaranteed to be an unsigned long long, so cast in printf
and use SCN macro in scanf.

int16_t isn't guaranteed to be a short, so use an SCN macro in scanf.

enums aren't guaranteed to have the same representation as int, so use
an intermediate variable.

+++ terps/plus/animations.c
@@ -56,7 +56,7 @@ void AddImageToBuffer(char *basename) {
     size_t len = strlen(basename) + 1;
     if (AnimationFilenames[ImgTail] != NULL)
         free(AnimationFilenames[ImgTail]);
-    AnimationFilenames[ImgTail] = MemAlloc((int)len);
+    AnimationFilenames[ImgTail] = MemAlloc(len);
     debug_print("AddImageToBuffer: Setting AnimationFilenames[%d] to %s\n", ImgTail, basename);
     memcpy(AnimationFilenames[ImgTail++], basename, len);
     AnimationFilenames[ImgTail] = NULL;
